+++ src/ListLiteral.php
@@ -73,10 +73,7 @@ final class ListLiteral extends AbstractLiteral
                 $elementType = $type;
                 continue;
             }
-            if ($elementType->equals($type)) {
-                continue;
-            }
-            $elementType = Type::any();
+            $elementType = $elementType->common($type);
         }
         return Type::listOf($elementType ?? Type::any());
     }